Added console_print files for color print. Fix requiring color issue for wild card.
