+++ include/console_print.h
@@ -1,0 +1,10 @@
+#ifndef __CONSOLE_PRINT_HEADER__
+#define __CONSOLE_PRINT_HEADER__
+
+#include <windows.h>
+#include <stdio.h>
+
+void print_warning(const char* string);
+void print_info(const char* string);
+
+#endif //__CONSOLE_PRINT_HEADER__
